backend/requirements.txt
psycopg2-binary==2.9.10
fastapi==0.115.12
fastapi-cli==0.0.7
passlib==1.7.4
python-jose==3.5.0
bcrypt==4.3.0